Fixed API report
