+++ workspaces/aws/plugins/amazon-ecs/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@alithya-oss/plugin-amazon-ecs",
-  "version": "0.4.1",
+  "version": "0.4.3",
   "repository": {
     "type": "git",
     "url": "github:awslabs/backstage-plugins-for-aws",
